Update function prototypes and the assign_specifier function options

+++ hex_fmt.c
@@ -10,7 +10,7 @@
  */
 char *hex_fmt(va_list args, FMT *fmt)
 {
-	long n = va_arg(args, int);
+	long n = va_arg(args, long);
 	int i = 1, hex_cap = fmt->type == 'X';
 	char *num = malloc(i + 1);
